+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Release (2025-XX-YY)
 - `stackitmarketplace`: [0.3.0](services/stackitmarketplace/CHANGELOG.md#v030-2025-04-04)
   - **Feature:** Add new `VendorProductId` attribute for subscription products
+- `resourcemanager`: [v0.4.0](services/resourcemanager/CHANGELOG.md#v040-2025-04-09)
+    - **Feature:** Add support for folder management
 
 ## Release (2025-03-27)
 - `alb`: [v0.1.0](services/alb/CHANGELOG.md#v010-2025-03-18)